Move logic into calculator.js
